+++ app/handlers/confirm_oauth_info.rb
@@ -4,7 +4,7 @@ class ConfirmOauthInfo
 
   uses_routine CreateEmailForUser
   uses_routine AgreeToTerms
-  uses_routine Newflow::StudentSignup::ActivateStudent
+  uses_routine StudentSignup::ActivateStudent
 
   paramify :signup do
     attribute :first_name
@@ -48,7 +48,7 @@ class ConfirmOauthInfo
     transfer_errors_from(@user, {type: :verbatim}, :fail_if_errors)
 
     agree_to_terms(@user) if options[:contracts_required] && signup_params.terms_accepted
-    run(Newflow::StudentSignup::ActivateStudent, @user)
+    run(StudentSignup::ActivateStudent, @user)
 
     outputs.user = @user
   end
