+++ spec/handlers/change_signup_email_spec.rb
