+++ spec/handlers/educator_signup/sheerid_webhook_spec.rb
@@ -1,8 +1,8 @@
 require 'rails_helper'
 require 'vcr_helper'
 
-#RSpec.describe Newflow::EducatorSignup::SheeridWebhook, type: :routine, vcr: VCR_OPTS do
-RSpec.describe Newflow::EducatorSignup::SheeridWebhook, type: :routine do
+#RSpec.describe EducatorSignup::SheeridWebhook, type: :routine, vcr: VCR_OPTS do
+RSpec.describe EducatorSignup::SheeridWebhook, type: :routine do
   let(:email_address)           { FactoryBot.create :email_address, :verified }
   let(:user)                    { email_address.user }
   let!(:school)                 {
