+++ spec/handlers/verify_email_by_code_spec.rb
@@ -26,7 +26,7 @@ describe VerifyEmailByCode, type: :handler, vcr: VCR_OPTS do
       end
 
       it 'activates the user' do
-        expect_any_instance_of(Newflow::StudentSignup::ActivateStudent).to receive(:exec).and_call_original
+        expect_any_instance_of(StudentSignup::ActivateStudent).to receive(:exec).and_call_original
         handler_call
         user.reload
         expect(user.state).to eq('activated')
@@ -42,7 +42,7 @@ describe VerifyEmailByCode, type: :handler, vcr: VCR_OPTS do
       let(:role) { User.roles[:instructor] }
 
       it 'verifies the email address found by the given code' do
-        allow_any_instance_of(Newflow::EducatorSignup::ActivateEducator).to receive(:exec).with(user: user)
+        allow_any_instance_of(EducatorSignup::ActivateEducator).to receive(:exec).with(user: user)
         expect(email.verified).to be(false)
         handler_call
         email.reload
@@ -50,12 +50,12 @@ describe VerifyEmailByCode, type: :handler, vcr: VCR_OPTS do
       end
 
       it 'calls EducatorSignup::ActivateEducator' do
-        expect_any_instance_of(Newflow::EducatorSignup::ActivateEducator).to receive(:exec).with(user: user)
+        expect_any_instance_of(EducatorSignup::ActivateEducator).to receive(:exec).with(user: user)
         handler_call
       end
 
       it 'outputs the user' do
-        allow_any_instance_of(Newflow::EducatorSignup::ActivateEducator).to receive(:exec).with(user: user)
+        allow_any_instance_of(EducatorSignup::ActivateEducator).to receive(:exec).with(user: user)
         outputs = handler_call.outputs
         expect(outputs.user).to eq(user)
       end
